+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 ## [[UnreleasedUniFFIVersion]] (backend crates: [[UnreleasedBackendVersion]]) - (_[[ReleaseDate]]_)
 
+### ⚠️ Breaking Changes ⚠️
+- UDL-based trait interfaces must now be wrapped with the `#[uniffi::trait_interface]` attribute.
+- Python: Trait interface implementations must now inherit from the trait base class.
+  This will look like `class PyTraitName(RustTraitName):`
+
 ### What's new?
 
 - Objects can export the `Ord` trait, allowing such objects to be ordered.
@@ -26,6 +31,14 @@
 - Towards `Enums` and `Records` getting methods:
   - Method metadata now carries a `MethodReceiver` for info about self.
   - In the templates, for `Callable`, `x.takes_self()` is replaced with `x.self_type().is_some()` etc.
+- Trait / Callback interface changes
+    - VTable fields are now: `free`, `clone`, followed by a field for each interface method.
+      Note That `free` is now at the start of the vtable rather than the end.
+    - Trait interface changes:
+        - Foreign handles must always have the lowest bit set
+        - Both Rust and foreign handles can now be passed across the FFI.
+          When Lifting/lowering trait interface handles, check if the handle was generated from Rust or the foreign side.
+    - See https://github.com/mozilla/uniffi-rs/pulls/2586 examples of how the builtin bindings here changed.
 
 [All changes in [[UnreleasedUniFFIVersion]]](https://github.com/mozilla/uniffi-rs/compare/v0.29.3...HEAD).